livy-api/py_requirements.txt
python-dotenv
azure-synapse-spark